Fix runtime error temporarily
There was a segmentation fault when a user clicked on "insert barcode" or insert other shapes in label editor. It was caused by using a method logicalDpiY() of and instance of QPaintDevice that was casted from QDesktopWidget to QPaintDevice. The problem was that QPaintDevice has a virtual method metric() that is implemented by QWidget - a parent of QDesktopWidget and a child of QPaintDevice.