Fix a runtime error
The runtime error occurred when a user chose to close the label editor window. It was caused by calling "delete this;" in LabelEditor::closeEvent() . Calling "delete this;" is not necessary as the label editor window is constructed with the Qt::WA_DeleteOnClose (Qt::WDestructiveClose in Qt3) flag.

Also port Qt::WDestructiveClose to Qt::WA_DeleteOnClose and Qt::WType_TopLevel to Qt::Window .