Fix a runtime error
Fix a runtime error caused by calling customHelpMenu() and helpMenu() in mainwindow.cpp. It is probably a bug in KDE/Qt because it showed up after an upgrade to a new Kubuntu and KDE/Qt version. These methods are deprecated anyway. So use the class KHelpMenu directly instead.

The positioning of an item when moving the item on the scene using a cursor works incorrectly. Resizing of items doesn't work. Painting artefacts are displayed when moving an item. Updating a cursor when it is over an item works incorrectly.