Fix positions of items
Before this fix if there were more than one item on the scene and a user started to move an item with the cursor the item jumped to other item's position. This happend becouse QGraphicsItem::isActive() works differently to Q3CanvasItem::isActive(). Altuogh I added TCanvasItem::setActive() and TCanvasItem::isActive() in an older commit the QGraphicsItem:isActive() was used when called. The solution was to change the names of my methods to TCanvasItem::setActiveItem() and TCanvasItem::isActiveItem().