Revert some changes made to rulers
The method KRuler::setMaxValue() was marked as deprecated in KDE 4.5 documentation and a warning "warning: ‘void KRuler::setMaxValue(int)’ is deprecated" is triggered when KBarcode's mycanvasview.cpp is compiled with KDE 4.8.3 . But this method is not marked as deprecated in KDE 4.8 documentation. Without calling setMaxValue() the rulers do not work correctly. Calling KRuler::setLength() does not solve this issue. So is the warning "deprecated" a bug here?