Fix the order of painting of items
Fix the order of painting of items in the preview and for the purpose of printing. Now the order of painting of overlapping (all actually) items (DocumentItem) is determined by 1) z value (the most important) 2) order of their addition to the scene (less important) 3) position (the least important). I had to add members (e.g. DocumentItem::additionOrder()) to a few classes becouse of that. Now the items in the preview correspond with the items in the scene view.

Before my attempt to set the painting order of document items, the ported items were drawn and printed in a random order (different every time the preview is updated or the printer was setup) which was very undesirable.