Fix a database related bug
The bug was: When a connection to a database is created and then the "Test settings" button in the configuration assistant is cliked the working default connection is removed and some of the database stuff doesn't work (e.g. loading "Producers" in the "Create a new label" dialog that is created when the label editor starts).