Display mispelled text
Display mispelled text of a selected text item(s) in Sonnet::Dialog. User can click on the Dialog's "Replace" button and then a replacement is done in the Dialog's buffer. But no replacement of text is applied to any text item.

If no mispelled text is found no dialog is displayed.