Fix batch assistant
The bug was: If a label was printed into a pdf file using Batch Printing -> Import variables and print -> Edit data manually / Import variables from a SQL table : Names of user variables (those in brackets: [...]) are printed into the pdf file instead of values for those variables even if values of the variables are inserted manually or loaded from a database and the values are displayed in the table widget of the batch assistant.

This bug was in Kbarcode for KDE3 too.