Improve batch assistant
old behaviour: Batch Printing -> Import variables and print -> Enter data manually / Import variables from a SQL table -> Next : If user edits variable values in the table widget by hand and then continues and then goes back all data in the table that user edited is lost and the table is blank or filled with original values from an sql table.

new behaviour: If a user visited the page with the table widget before and the table widget has rows a dialog "What to do with already inserted data?" is displayed with buttons "Keep", "Replace", "Append".