Add support for old versions of BWIPP
BWIPP stands for Barcode Writer in Pure PostScript.
Check content of the file barcode.ps for it's version
at runtime (when application starts)
and adjust the postscript command according to it.

On 2010-01-05 Terry Burton converted BWIPP to a set of PostScript
named resources rather than a procedure list.
So the related postscript command changed as well.