Add configuration of Barcode Writer In Pure Postscript
Add options to the configuration dialog
that user can use to choose which version of Barcode Writer In Pure Postscript
located on a local system will be used by KBarcode-kde4 to generate barcodes.
Add the following options to the configuration dialog:
1) Automatic - Let KBarcode-kde4 choose the newest version
of Barcode Writer In Pure Postscript automatically.
Either the version installed with the package libpostscriptbarcode
or the one shipped with KBarcode-kde4 is used, depending on
which one is the most up-to-date.
2) Use the version installed with the package libpostscriptbarcode
into /usr/share/libpostscriptbarcode/barcode.ps .
This option is disabled if the file is not found.
3) Use the version shipped with KBarcode-kde4.
The library is in the file barcode.ps which is installed
into KBarcode-kde4's data directory.
4) Set another version located in a file on the local system.