Remove the option "Start with an empty label"
Remove the checkbox option "Start with an empty label"
from the "new label" dialog when the label editor is started.
Remove it becouse it is confusing to users
and it doesn't seem to be useful at all.
The decision has been made by me (František (Fanda) Brožka)
and Gianluca Montecchi.

Email from Dominik Seichter on this matter:
I think the "Start with an empy label" checkbox is some kind of workaround
to start the KBarcode label designer without having to choose a label
definition. The only use that comes to my mind is that you can then open
another existing label definition (File -> Open) or access the
configuration dialogs etc. .
It seems like I chose pretty bad wording for this option and that there
could be better ways to achieve the same feature (like allow to open files
directly from the open dialog, with also access to recent files maybe).