Fix line styles' icon size
Fix line styles' icon size. The bug was reported by Gianluca Montecchi.

From the message by Gianluca Montecchi:
"Anyway, I have an observation about the label editor:
when you insert something that has a border (for example a rectangle),
in the dialog to set the properties of the object,
in tab "Border" there is the option to set the line style.
The combo to choose the "Line style" shows too small example
of the actual style of the border,
which can render hard to select the correct style for the border line.
This is for all the object that has the border and for the "line" object."